backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
playwright>=1.40.0
sqlalchemy>=2.0.0
loguru>=0.7.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
akshare>=1.11.0

# Mira library (install from local path)
# Run: pip install -e ../mira

